+++ src/Fp.Fs/FormatSingleProcessorFsWrapper.cs
@@ -24,7 +24,7 @@ public class FormatSingleProcessorFsWrapper : FsProcessor
     {
         var input = OpenMainFile();
         _baseProcessor.Prepare(input, Name, Configuration);
-        return _baseProcessor.TryProcess(out Data? data) ? new[] { data } : Array.Empty<Data>();
+        return _baseProcessor.TryProcess(out Data? data) ? new[] { data! } : Array.Empty<Data>();
     }
 }