+++ src/Fp.Fs/FsProcessor.Run.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using static Fp.FsProcessor;
 
 namespace Fp;
 
@@ -82,47 +81,3 @@ public partial class FsProcessor
     public static void Run<T>(IList<string>? args, string name, string description, params string?[] extensions) where T : FsProcessor, new() =>
         Run(null, args, new GenericNewFsProcessorFactory<T>(new FsProcessorInfo(name, description, description, extensions)));
 }
-
-public partial class Scripting
-{
-    /// <summary>
-    /// Processes using direct function.
-    /// </summary>
-    /// <param name="func">Function or delegate run per file.</param>
-    /// <param name="args">Arguments. If null, only register processor.</param>
-    /// <param name="info">Processor info.</param>
-    public static void fp(Action func, IList<string>? args, FsProcessorInfo? info = null) =>
-        Run(func, args, info);
-
-    /// <summary>
-    /// Processes using direct function.
-    /// </summary>
-    /// <param name="func">Function or delegate run per file.</param>
-    /// <param name="args">Arguments. If null, only register processor.</param>
-    /// <param name="name">Processor name.</param>
-    /// <param name="description">Processor description.</param>
-    /// <param name="extensions">Processor extensions.</param>
-    public static void fp(Action func, IList<string>? args, string name, string description, params string?[] extensions) =>
-        Run(func, args, name, description, extensions);
-
-    /// <summary>
-    /// Processes using segmented function.
-    /// </summary>
-    /// <param name="func">Function that returns enumerable (segmented processing enumerator).</param>
-    /// <param name="args">Arguments. If null, only register processor.</param>
-    /// <param name="info">Processor info.</param>
-    public static void fp(Func<IEnumerable<Data>> func, IList<string>? args, FsProcessorInfo? info = null) =>
-        Run(func, args, info);
-
-    /// <summary>
-    /// Processes using segmented function.
-    /// </summary>
-    /// <param name="func">Function that returns enumerable (segmented processing enumerator).</param>
-    /// <param name="args">Arguments. If null, only register processor.</param>
-    /// <param name="name">Processor name.</param>
-    /// <param name="description">Processor description.</param>
-    /// <param name="extensions">Processor extensions.</param>
-    public static void fp(Func<IEnumerable<Data>> func, IList<string>? args, string name, string description, params string?[] extensions) =>
-        Run(func, args, name, description, extensions);
-}
-// ReSharper restore InconsistentNaming