+++ src/Fp/FormatProcessor.cs
@@ -11,36 +11,56 @@ public class FormatProcessor : Processor
     /// <summary>
     /// Current file name.
     /// </summary>
-    public string Name { get; set; } = "";
+    public string InputFile { get; set; } = "";
+
+    /// <summary>
+    /// Current file name.
+    /// </summary>
+    public string Name => Path.GetFileName(InputFile);
+
+    /// <summary>
+    /// Current file name without extension.
+    /// </summary>
+    public string NameNoExt => Path.GetFileNameWithoutExtension(InputFile);
+
+    /// <summary>
+    /// Current file name.
+    /// </summary>
+    public FpPath NamePath => FpPath.GetFromString(Name) ?? throw new InvalidOperationException();
+
+    /// <summary>
+    /// Current file name without extension.
+    /// </summary>
+    public FpPath NamePathNoExt => FpPath.GetFromString(NameNoExt) ?? throw new InvalidOperationException();
 
     /// <summary>
     /// Prepares critical state for operation.
     /// </summary>
     /// <param name="stream">Source stream.</param>
-    /// <param name="name">Input name.</param>
+    /// <param name="inputFile">Input name.</param>
     /// <param name="configuration">Additional configuration object.</param>
-    public void Prepare(Stream stream, string name, ProcessorConfiguration? configuration = null)
+    public void Prepare(Stream stream, string inputFile, ProcessorConfiguration? configuration = null)
     {
         Prepare(configuration);
         UseStream(stream);
-        Name = name;
+        InputFile = inputFile;
     }
 
     /// <summary>
     /// Prepares critical state for operation.
     /// </summary>
     /// <param name="memory">Source memory.</param>
-    /// <param name="name">Input name.</param>
+    /// <param name="inputFile">Input name.</param>
     /// <param name="configuration">Additional configuration object.</param>
-    public void Prepare(ReadOnlyMemory<byte> memory, string name, ProcessorConfiguration? configuration = null) => Prepare(new MStream(memory), name, configuration);
+    public void Prepare(ReadOnlyMemory<byte> memory, string inputFile, ProcessorConfiguration? configuration = null) => Prepare(new MStream(memory), inputFile, configuration);
 
     /// <summary>
     /// Prepares critical state for operation.
     /// </summary>
     /// <param name="data">Source data.</param>
-    /// <param name="name">Input name.</param>
+    /// <param name="inputFile">Input name.</param>
     /// <param name="configuration">Additional configuration object.</param>
-    public void Prepare(BufferData<byte> data, string name, ProcessorConfiguration? configuration = null) => Prepare(new MStream(data.Buffer), name, configuration);
+    public void Prepare(BufferData<byte> data, string inputFile, ProcessorConfiguration? configuration = null) => Prepare(new MStream(data.Buffer), inputFile, configuration);
 
     /// <summary>
     /// Prepares critical state for operation.