+++ src/Fp/Scripting.cs
@@ -1,8 +1,0 @@
-namespace Fp;
-
-/// <summary>
-/// Provides scripting-related functions/properties.
-/// </summary>
-public static partial class Scripting
-{
-}