+++ src/Fp.Plus/CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.24.0]
+## [Unreleased]
+### Changed
+- Updated `Fp` to `0.25.0`
+
+## [0.24.0] 2022-1-16
 ### Changed
 - Updated `Fp` to `0.24.0`
--
+
 ## [0.23.1] 2021-12-27
 ### Changed
 - Updated `Fp` to `0.23.1`